+++ pkg/watcher/certificate/certmanager/certificate.go
@@ -1,4 +1,4 @@
-package cert_manager
+package certmanager
 
 import (
 	"context"
